Bug# 13011344 - PURGE VIEW DOESN"T NEED TO BE IN THE VIEW LIST
There is no need for the purge view to be on the read view list. The read
view list is required only for purge. So that it can clone the oldest view.
This change allows the release of the trx_sys_t::mutex just after we've cloned
the view. Before this fix we had to keep the mutex locked right up to the
point where we added the purge view to the read view list. This was required
to maintain the order of the read view list.

Approved by Jimmy rb://770