Bug# 13011596 - THE STATS COUNTING IS EXPENSIVE DUE TO THE MUTEX WHEN DOING IO
The mutex calls that protect the IO counters are expensive. This is with a load
with checksums=off. Otherwise the checksum code dominates. 
      
Make the monitor inc/dec atomic. The accuracy of the min and max values is not
100% due to an inherent race. To use atomics add another condition that
UNIV_WORD_SIZE == 8 because the monitor counters use ib_int64_t. 
      
Approved by Jimmy rb://751