BUG#12987771: REPLICATED STATEMENTS ARE DUPLICATED IN THE SLAVE GENERAL LOG
This is a regression that was introduced by WL 5706. In that
worklog some changes were made in order to avoid that the server
would log to general, slow and binary log passwords in clear
text. As part of that work, a couple of general_log_write calls
were deployed in mysql_parse.
     
Now, the slave thread, when applying a query_log_event, calls
mysql_parse as part of the apply execution. This makes the slave
thread to write the statement twice to the log (an extra time at
mysql_parse).
     
We fix this by checking if it is the slave thread that is calling
mysql_parse. If it is, the decision of what to log to the error
log is taken inside Query_log_event::do_apply_event (as it was
before), thus general_log_write is not called at this point.
      
A side change is that slave thread was ignoring the opt_log_raw
value. When set, for user sessions, the server logs the raw
statement, disregarding if it is rewritten or not. We make the
slave thread behave the same.

[ NOTE: merged / cherrypicked approved cset that had been committed
  against mysql-trunk ]