BUG#12837714: ADDITIONAL NULL IN 5.6 ON GROUPED SELECT
The bug was that Cached_item_field::cmp() did not do proper 
NULL checking: if both buff and the field values were NULL, 
the result was based on bitwise field comparison. Since a 
field with NULL value may contain random bits, buff and 
field were considered different in this case.

The fix is to handle the case where both are NULL: the 
returned value from the function is now that they are equal.

Note: there are four more Cached_item_* classes: For the int,
real and decimal versions it is OK to compare cached value 
to the new value because val_int/decimal/real() return 0 if 
the value is NULL. For the string cached item, we get the 
same issue as with Cached_item_field but for this class the 
case where cached value and new value are both NULL is already 
handled.