Adding regression test for the following commit:
"4562 Jan Wedvik 2011-09-29
This patch will make the SPJ block fetch all rows for some non-root index scans
in one batch rather than two.
This will happen if the first batch reads from a subset of the fragments and
receive few rows. The SPJ block will then
try to read from the remaining fragments before finishing the batch.
This is especially useful when doing bushy scans. If there are more branches in
the bushy scan, then these will have to be
repeated for each batch of the current scan."