BUG#11747866:REPLICATE-REWRITE-DB DOESN'T WORK WITH SINGLE CHARACTER DATABASE NAMES
Problem: replicate-rewrite-db option doesnot work with single character database names.
The last character was being trimmed from the "from" database name.

Background: Since the loop handling the option, has to handle invalid arguments for the
option,while stripping the whitespaces, it was stripping one character from the "from" 
database name. In case where there was only one character, this threw an error as
"empty from database name".

Fix. Revamped the loop handling this option. Made sure that the whitespaces are
suitably trimmed and the correct "to" and "from" databases names are passed to the 
function handling the replicate_rewrite_db functionality.