Bug #12330344 CRASH AND/OR VALGRIND ERRORS IN FREE_IO_CACHE WITH JOIN, VIEW, PARTITIONED TABLE
Problem: Server crashes when executing query using a TEMPTABLE view
with a subselect using a partitioned table.

In prune_partitions(), the memory root is substituted for a memory
root with a life span as long as the prune_partions() call. During
optimization of functions, subselects may allocate memory that is
expected to have the same life span as the subselect. Due to the new
memory root, this memory is freed at the end of prune_partitions() and
still referred to by the subselect and attempted used during query
execution.

Fix: Switch memory root back to the normal root before calling
Item_func::select_optimize().