Bug#13007154: Crash in keys_to_use_for_scanning with ORDER BY                            and PARTITIONING
The server could crash on a query that called
ha_partition::keys_to_use_for_scanning when there was no partitions
used.Due to the first used partition would then be out of bounds.

The solution was to always use the first partitions handler since the
function will return the same key_map for every partition, since they
have the same engine and the same structure.

This bug was introduced as a regression in WL#5217.