Bug#12960284 - RPL.RPL_PARALLEL_SWITCH_SEQUENTIAL TEST FAILED IN VALGRIND RUN ON
There was a timeout at the very end of the test.
Details of printout upon the failure does not reveal any issue with the text nor
the server code. It's just slow and is used to pass on the same valgrind but the run
time is apparently big (over 1 000 000 msec) -
  rpl.rpl_parallel_switch_sequential 'mix' w2 [ pass ]  1068101

So it's fixed with decreasing load size when the test runs on valgrind.