This commit concerns SPJ (i.e. pushed queries).
The commit fixes two errors that may happend when there is a pruned child
scan operation:
1. The api did not set SIP_PRUNE_PARAMS as it should.
2. There was an error in Dbspj::parseScanIndex().  When it called 
Dbspj::expand() to process prune keys, it used to start reading
query parameters from after whatever parseDA() had consumed. Now it starts 
from the beginning of the query parameters (for that operation). In other 
words, the parameter values are only sent once, and these values are used for 
building both the scan bounds and the prune key.