Bug#12664936: Same query executed as where subquery gives different               results on in() compare #2
Step 1 of 2 - Refactoring

This is a local refactoring done before the actual bugfix.
It makes get_best_combination() a public member function of class
JOIN, and splits the main work over a few smaller functions.
It also changes the order of some evaluations so that compiled
semi-join information is now available in the JOIN_TAB objects at
an earlier stage.
get_best_combination() should actually be private, because it is
only called from make_join_statistics(), which is a free function
that should also be a private member function of class JOIN.

sql/sql_select.cc
  get_best_combination() is refactored according to pseudo code:
   - Fix semi-join strategies after table order selection.
   - Create the new JOIN_TAB array.
   - A loop that initializes this JOIN_TAB array.
   - A function set_semijoin_info() that sets semi-join information
     for the selected table order in the JOIN_TAB array.
   - A function set_access_methods() that sets up access methods
     based on key information.
   - Update dependencies between tables based on the selected
     key access patterns.

sql/sql_select.h
  Defined get_best_combination() as public member function of JOIN.
  Created set_semijoin_info() and set_access_methods() as private
  member functions of class JOIN.