Bug#113106350: crash in memcpy from join_cache::write_record_data
This problem may happen if we combine the Duplicate Weedout semi-join
strategy with join buffering, and the outer table is a derived table.

The problem is that Duplicate Weedout needs the row id of all involved
tables, hence the table is instructed to save the row id.
Join buffering needs to save and restore the rowid into a buffer,
for that purpose an st_cache_field object is allocated. This object
contains a pointer to the row id buffer, but at the time of object
creation, the table (being a derived table) has not yet been opened
and the buffer pointer is NULL.

The fix is to save a reference to the st_cache_field object in the
corresponding JOIN_TAB and bind it with the buffer after the derived
table has been materialized.

mysql-test/t/derived.test
  Added test case for bug#113106350.

mysql-test/r/derived.result
  Added test case results for bug#113106350.

sql/sql_join_cache.cc
  Save reference to rowid copy object when setting up fields to copy.

sql/sql_select.cc
  Bind the rowid buffer of a TABLE object to a copy object that is
  used to copy a rowid. Needed when temporary tables are opened
  as part of execution.
  
sql/sql_select.h
  Added reference to copy object used to copy rowid in JOIN_TAB.
  Added a bind_buffer() access function for st_cache_field.