Bug#13113222 RQG_SIGNAL_RESIGNAL FAILED WITH ASSERTION.
The triggered assert checks that a matching condition is
found inside the diagnostics area when a handler is activated.
It was triggered in this case if the diagnostics area was
full when the condition was reported (e.g. the area already
contains max_error_count conditions). This meant that the
diagnostics area did not contain a matching condition when
the handler code tried to look for it.

This was a regression introduced by WL#5986.

This patch fixes the problem by temporarily making a condition
in such cases so it can be handled properly.

Test case added to sp-error.test.