Bug#13326965 THD->CURRENT_UTIME() CALLED WHEN NOT NEEDED IN              LOG_SLOW_STATEMENT()
The problem was that thd->current_utime() was called before
writing to the slow log even if no log entries ended up being
written. This is unneccessary and potentially a performance
problem (thd->current_utime() -> my_micro_time() -> 
__gettimeofday()).

This patch fixes the problem by not calling thd->current_utime()
before it is known that a log entry will in fact be written.

The patch doesn't contain a test case since it is hard to test
for performance improvements in our test framework.