Bug #13098726 MY_TIME.C:786: CALC_DAYNR: ASSERTION `DELSUM+(INT) Y/4-TEMP > 0' FAILED
Problem: The date 0000-01-00 triggers an assertion guarding against
invalid results in calc_daynr().

An initial check in calc_daynr() returns day number 0 for all dates in
month 0, year 0. The assertion at the end of the function states that
other return values should be greater than 0. However, because of the
way day numbers are calculated, the fictitious date 0000-01-00 is also
represented as day number 0, and this wrongly triggers the assertion.

In most cases, the NO_ZERO_IN_DATE SQL mode causes the problematic
date to be handled outside calc_daynr(), but not all time functions
check for such fictitious dates.

Fix: Change assertion so that it states that return values should be
greater than or equal to zero.