ndb - rename class Rope to LocalRope
No functional changes.  

Only renaming to emphasize that the class shares some of the
behaviour with other "Local" classes in ndb.  For example,
intended for defining object as local variables where constructor
copies some non-local object and the destructor copies back a
possible changed state to the non-local object on deletion.

So beware, dont change or destroy the non-local object while the
local copy is alive.