Bug 12980094 - ASSERTION IN INNODB DETECTED IN RQG_PARTITION_DDL
This bug is a debug code bug, it is only possible on platforms that don't
support atomics.

Release the rw_lock before freeing the underlying mutex. This fixes a race
condition in rw_lock_list_print_info(), where the underlying mutex was freed
but the rw_lock was still in the rw_lock_list.

Approved by Jimmy Yang over IM.