Fix for Bug#13355055 CLUSTER INTERNALS FAILS TO TERMINATE BATCH AT MAX 'BATCHBYTESIZE'
There is a mixup in whether 'ScanFragConf::total_len' and TupKeyConf::readLength
specifies the returned length in #bytes or #Uint32 - It is specified in #Uint32, 
but interpreted as being number of bytes by the LQH and SPJ block.

Correct this such that returned length is multiplied by sizeof(Uint32) before 
it is added to 'm_curr_batch_size_bytes'. ScanRecord::check_scan_batch_completed()
will later use the aggregated 'm_curr_batch_size_bytes' to determine when the
batch has been filled up.

When later returning 'ScanFragCOnf::total_length' which still is in number 
of Uint32's, we have to convert back from #bytes to #Uint32s.