WL#5559 -  Factor tmp table out of optimizer.   Patch 2: Remove curr_join references from JOIN::exec
Most references in JOIN::exec to member fields are indirect
through local variable curr_join, which may either refer to
the original JOIN object or a temporary copy of it.  This
complicates the code significantly.

In order to simplify the code and ease further refactoring
of JOIN::exec, this patch removes the need for curr_join by, 
when a temporary copy is used, executing the code directly 
on the temporary object.

This is achieved by creating a private method, 
JOIN::execute(JOIN *parent) to be called from JOIN::exec(),
and move all code after creation of the first temporary table
to this method. If a temporary copy is needed, then 
tmp_join->execute(this) is called. Otherwise,
this->execute(NULL).

The code makes sure that necessary execution data is copied
from the temporary JOIN object and back to its parent in order
to be able to reuse it for later executions as before.

A limited number of example queries has been used to verify that
there is no performance regression.  Also, using gprof it has been
verified that the functions previously called from JOIN::exec, 
is still called the same number of times with the new approach.