Avoid problems with support for binlog-row-image
The binlog-row-image option allows row events to be generated either
as FULL, MINIMAL, or NOBLOB.  

FULL causes all columns, not just those modified, to be logged in 
each event AND it causes update and delete events to have full 
BEFORE images.

MINIMAL logs only modified columns, and update and delete events do
not need before images if the table has a primary key.

NOBLOB is similar to FULL, except unmodified BLOB columns are not 
logged.

The default value is FULL.

The FULL option is implemented by setting all columns in the write
set when performing an Insert, Update or Delete operation.

The MINIMAL option avoids this, and records only the modified columns
in the Binlog, and skips before images if not required.

Ndb already supports similar functionality via the 
--ndb-log-updated-only server option and per-table binlog flags.  This
is implemented in trigger logic in the data nodes, and affects non-MySQLD
sourced modifications as well.

The part of the FULL implementation which causes operation's write
sets to be fully set breaks the ndb implementation of ndb-log-updated-only
as all columns are always written to.

For this reason, the HTON_NO_BINLOG_ROW_OPT flag is set by Ndb to 
avoid the FULL option having this effect.

The FULL, MINIMAL or NOBLOB option still affects how the Ndb Binlog
Injector records events, in conjuntion with --ndb-log-updated-only.
The default value, FULL, behaves as it did prior to mysql-trunk-cluster.

When MINIMAL is passed, Delete and Update events do not need to have
before images, saving space in the Binlog.

The --binlog-row-image option can be set per-server or per-session.
The Ndb engine uses the per-server setting and ignores per-session
settings. 

This patch results in the following testcases passing :
ndb_binlog_variants, ndb_rpl_conflict_max, 
ndb_rpl_conflict_old, ndb_rpl_2innodb, ndb_rpl_basic, 
ndb_rpl_idempotent, ndb_rpl_logging