Fix bug#13363044 FAILING ASSERTION: ERR == DB_SUCCESS IN FILE ROW0VERS.C LINE 144
Checking for DB_SUCCES alone is not sufficient. Not being able to find a
previous version of the record is a perfectly valid state. This can happen
because the transaction that created the record has committed and is not
in purge's view. The caller of row_vers_impl_x_locked_low() has to check
again whether the transaction is active or not, which it does.

Approved by Jimmy over IM.