ndb - adding a schema file id hash table for DictObject
And removing the current separate hash tables for File,
Filegroup, and HashMapRecord.

The id hash for DictObject replaces the current hash tables
for File, Filegroup, HashMap and will also be used by Table
and Index.  Note that temporary table objects created by
alter table will be added in id hash table but not in
name hash table, this since the way to reach tablerecords
will be through dictobject and schema file id, not as now
assume tableid is ptr-i-value in tablerecordpool.