Bug#12578908: SELECT SQL_BUFFER_RESULT OUTPUTS TOO MANY ROWS               WHEN GROUP IS CONSTANT
For all but simple grouped queries, temporary tables are used to
resolve grouping. In these cases, the list of grouping fields is
stored in the temporary table and grouping is resolved
there (e.g. by adding a unique constraint on the involved
fields). Because of this, grouping is not done when inserting
rows into the temporary table.

In the case where a group clause may be optimized away, grouping
does not have to be resolved using a temporary table. However, if
a temporary table is explicitly requested (e.g. because the
SQL_BUFFER_RESULT hint is used, or the statement is
INSERT...SELECT), a temporary table is used anyway. In this case,
the temporary table is created with an empty group list (because
the group clause was optimized away) and it will therefore not
create groups. At the same time, make_simple_join() assumes that
since a temporary table is used, grouping will be taken care of
and sets JOIN::group to false. 

The fix is in make_simple_join(), and is to not set JOIN::group
to false in the unlikely event that a temporary table is used for
a query where the group clause was optimized away. With this
change, end_{send|write}_group() is used when reading from the
temporary table. This is the same function as is used when a
temporary table is not explicitly requested.