Fix: SPJ may leak RowMaps
There is a minor memory leak of 'struct RowMap' objects.
As the RowMap was only ::init()'ed when Dbspj::releaseNodeRows() has
released all the mapped rows, its reference was effectively lost
at that point and another RowMap will be allocated if needed

... NOTE: 'leak' in this context is not a true memory leak
as all memory is managed within the request and released
when ::cleanup() removes all objects related to this request.