Bug#13371000 - ADAPTIVE LRU FOR COMPRESSED PAGES BROKEN BY RECENT LRU CHANGES
rb://812
approved by: Marko Makela

When calling buf_flush_LRU_list_batch() use appropriate LRU list
i.e.: either the regular LRU list or the unzip_LRU for cleaning and
eviction. If it is decided to use the unzip_LRU then we can simply
skip the flushing part because we are going to keep the compressed
frames.