Bug#13410464: FOR START SLAVE .. PASSWORD=<PASSWORD>, THE PASSWORD               IS LOGGED IN PLAIN TEXT
The server was not rewriting the statement START SLAVE when it
was issued along with the PASSWORD instruction. Thence, the
password is written in clear text to the general log.

To fix this, we register the SQLCOM_SLAVE_START as a command that
should be rewritten (if the command contained indeed the PASSWORD
token).