Bug#11748572:ALLOCATING A LARGE QUERY CACHE IS NOT DETERMINISTIC
Setting query_cache_size to larger values might fail depending on the memory 
pressure being put on the system. This can be seen on pushbuild as the test 
case query_cache_size_basic tries to allocate a +3GB query cache, which 
succeeds in some machines and fails in others.

So this part of the code where the test case tries to allocate +3GB query cache has been 
disabled for now to get the test running on pb2.