Fix Bug#12939557 VALGRIND:VEX AMD64->IR: UNHANDLED INSTRUCTION BYTES: 0XF2 0X48 0XF 0X38 0XF0 0XA
Emit an error from CMake if WITH_VALGRIND is turned ON (it is OFF by
default) and if Valgrind headers are not found. Before this patch CMake
silently switched WITH_VALGRIND to OFF and continued if the needed
headers were not found.                                                   

We have experienced a few time-comsuming issues with people turning ON
WITH_VALGRIND and then not getting InnoDB Valgrind instrumentation
enabled because CMake couldn't find the headers and disabled WITH_VALGRIND.